simulation: rename class NodeState to State
Rename class `infection.node.NodeState` to `infection.node.State`.
There is no other state we care about at the moment, so we have no
reason to keep this discrepancy in the class naming convention.

If we will care of other states in the future, we will still be able to
use the `import-as` constuct where needed.