simulation: add options for directories and saving
There are now two ways for passing the graph:
- with option `-G` read graph from file: this replaces the previous
optional last argument GRAPH-FILE;
- with option `-g` read graph with a given UID from `./graphs`
directory, or any other directory specified with `--graph-dir`.

With option `--save` the simulation is saved in `./simulations`; this
directory can be changed with option `--simulation-dir`.

With `--save`, comes the optional `--verbose`, which prints the
simulation directory and UID in a pretty way; otherwise only the
simulation UID is printed.