util: add function `string_to_linspace`
Function `util.string_to_linspace` generates a list of evenly spaced
float numbers from a string in the format "[START[,STOP[,NUM]]".

If only START is given, the list is [START]; if also STOP is given, the
list contains NUM evenly spaced numbers from START to STOP; the default
for NUM is 11.

This function is mainly a wrapper of `numpy.linspace`, which requires at
least 2 arguments.